# Performance: Use subclass for parsing EXIF data