The old method (using dict() and _fixup_dict) create clones from IFD_v1, but it read all items when iterating, losing all benefits from the lazyness of IFD.
The new method use a custom subclass that try to prevent all iterations as much as possible, keeping all tags undecoded.

The return value isn't a dict anymore, but a subclass of ImageFileDirectory_v1 that doesn't support tov2(). All decoded tags are formatted as before.

With a test image, I'm going from:
```py
%timeit Image.open("Tests/images/exif-dpi-zerodivision.jpg") 1000 loops, best of 5: 485 µs per loop
```
to
```py
%timeit Image.open("Tests/images/exif-dpi-zerodivision.jpg") 1000 loops, best of 5: 257 µs per loop
```

With a more common JPEG image from my Panasonic TZ10 camera, I'm going from:
```py
%timeit Image.open("/home/photos/P1040526.JPG") 1000 loops, best of 5: 3.24 ms per loop
```
to
```py
%timeit Image.open("/home/photos/P1040526.JPG") 1000 loops, best of 5: 871 µs per loop
```

I hope that even if the return type of `_getexif` is not the same, its compatibility with the previous format will make it merged.