@radarhere 2019-02-18
Would you be able to add tests?