@Glandos 2019-02-18
Since no external interfaces were modified, I don't see the point to add any tests.

Of course, we could imagine tests for this new class, but they should already be covered by the existing EXIF tests. In fact, existing tests helped me **a lot** to avoid making useless mistakes.

Could you be more precise on which part of the code could be tested?