@radarhere 2019-02-19
I would think that adding a new class constitutes a new interface, and so I would suggest adding independent tests for it. Feel free to disagree, but on the other hand, if you think that this class should not be used by end users, then why have the `ExifImageFileDirectory.to_v2` method?

Normally, we would have an automatic report here about the change in code coverage as a result of this PR, so I might also be able to talk about keeping code coverage up generally, but #3641 is preventing that report.