@Glandos 2019-02-20
My goal was to return lazy version of the current dict in `_getexif`, with some optimized version of `ImageFileDirectory_v1` that is currently used.

So _normally_, the output class should be transparent. However, the `to_v2` override is here to prevent misuse. I can also simply delete the method, so that it is impossible to call.

I don't know if it's possible (or even desirable) to hide this class declaration.

In case you really want tests for that new class, I will need some help, because there is currently no test for `ImageFileDirectory_v1`, there is only tests for `_getexif` that uses this class.