@Glandos 2019-04-18
It seems that the merge of #3625 make this request invalid, but not irrelevant.

I still have the following performance, when opening 1000 times the same JPEG image with a lot of EXIF tags produced by a camera:

![Test](https://screenshotscdn.firefoxusercontent.com/images/bb1aa33f-44eb-45d1-97ea-a05f5572bcc4.png)

It seems that I can adapt this request to the new layout. Do you have any objections on that before I spend some of my spare time on it?