@radarhere 2019-08-18
The image link from your last comment no longer works.

I've created PR #4031 as a new version of this PR - see what you think.