@Glandos 2019-08-20
> The image link from your last comment no longer works.
I really dislike Github for their obscure retention policy.

> I've created PR #4031 as a new version of this PR - see what you think.
Many, many thanks for taking care of that. It is really different from mine, but the code has changed too, and I think it does the job.